+++ GRDB/Core/DatabaseReader.swift
@@ -149,6 +149,27 @@ public protocol DatabaseReader : class {
     
     /// Remove a collation.
     func remove(collation: DatabaseCollation)
+    
+    // MARK: - Value Observation
+    
+    /// Starts a value observation.
+    ///
+    /// You should use the `ValueObservation.start(in:onError:onChange:)`
+    /// method instead.
+    ///
+    /// - parameter observation: the stared observation
+    /// - parameter onError: a closure that is provided by eventual errors that happen
+    /// during observation
+    /// - parameter onChange: a closure that is provided fresh values
+    /// - returns: a TransactionObserver
+    func add<Reducer: ValueReducer>(
+        observation: ValueObservation<Reducer>,
+        onError: ((Error) -> Void)?,
+        onChange: @escaping (Reducer.Value) -> Void)
+        throws -> TransactionObserver
+    
+    /// Remove a transaction observer.
+    func remove(transactionObserver: TransactionObserver)
 }
 
 extension DatabaseReader {
@@ -227,5 +248,22 @@ public final class AnyDatabaseReader : DatabaseReader {
     /// :nodoc:
     public func remove(collation: DatabaseCollation) {
         base.remove(collation: collation)
+    }
+    
+    // MARK: - Value Observation
+    
+    /// :nodoc:
+    public func add<Reducer: ValueReducer>(
+        observation: ValueObservation<Reducer>,
+        onError: ((Error) -> Void)?,
+        onChange: @escaping (Reducer.Value) -> Void)
+        throws -> TransactionObserver
+    {
+        return try base.add(observation: observation, onError: onError, onChange: onChange)
+    }
+    
+    /// :nodoc:
+    public func remove(transactionObserver: TransactionObserver) {
+        base.remove(transactionObserver: transactionObserver)
     }
 }