+++ Tests/GRDBTests/GRDBTestCase.swift
@@ -166,3 +166,11 @@ class GRDBTestCase: XCTestCase {
         }
     }
 }
+
+#if !swift(>=4.2)
+extension Sequence {
+    func allSatisfy(_ predicate: (Self.Element) throws -> Bool) rethrows -> Bool {
+        return try !contains(where: { try !predicate($0) })
+    }
+}
+#endif